requirements.txt
numpy==1.18.5
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.4
paramiko==2.7.1
Pillow==7.1.2
psutil==5.7.0
PyAutoGUI==0.9.50
PyInstaller==3.6
python-dateutil==2.8.0
PyYAML==5.3.1
scapy==2.4.3
scapy-python3==0.26
scipy==1.4.1
uiautomation==1.1.15
urllib3==1.25.9
WMI==1.5.1
xlrd==1.2.0
numba==0.45.1
chardet==3.0.4
#latest llvmlite version has issue with numba
llvmlite==0.29.0
mss==5.1.0
pyUserInput==0.1.10
jinja2==2.10.1
markupsafe==1.1.1
requests==2.24.0
certifi==2020.6.20
idna==2.10
#pyqt5 need add install location to system path
pyqt5==5.13.0
pyserial==3.4
winshell==0.6
selenium==3.141.0
pywifi==1.1.12
wakeonlan==2.0.1